+++ sdk/core/System.ClientModel/tests/Pipeline/RequestRetryPolicyTests.cs
@@ -1,10 +1,10 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using ClientModel.TestHelpers.Mocks;
 using NUnit.Framework;
 using System.ClientModel.Primitives;
 using System.Collections.Generic;
-using TestHelpers.Internal;
 
 namespace System.ClientModel.Tests;
